# chore(warehouse): event schema performance improvements