--- a/event-schema/event_schema_suite_test.go
+++ b/event-schema/event_schema_suite_test.go
@@ -1,0 +1,13 @@
+package event_schema
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestEventSchemas(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Event Schemas Suite")
+}