… event-schema/event_schema.go
… event-schema/testdata/test_input.json