# Description
This PR limits the frequency counters we are creating and storing for event models and schema versions.


## Notion Ticket
[Notion - PR](https://www.notion.so/rudderstacks/Event-Schema-OOM-Kill-Issue-64b6d7384d704f1a987696012c6fd86a)


## Security

- [x] The code changed/added as part of this pull request won't create any security issues with how the software is being used.


## Benchmark Result
```
name                      old time/op    new time/op    delta
EventSchemaHandleEvent-8     2.55s ± 0%     2.38s ± 0%   ~     (p=1.000 n=1+1)

name                      old alloc/op   new alloc/op   delta
EventSchemaHandleEvent-8    8.52MB ± 0%    4.46MB ± 0%   ~     (p=1.000 n=1+1)

name                      old allocs/op  new allocs/op  delta
EventSchemaHandleEvent-8     55.1k ± 0%     31.6k ± 0%   ~     (p=1.000 n=1+1)
```